Emotion scores from 2016 presidential debate

A data frame containing predicted emotion scores from the first 2016 US presidential debate. Each row is the score for each emotion type for either Trump or Clinton in each frame of the video. Emotion scores calculated using the Microsoft Emotion API.

emotion

Format

A data frame with 128,688 rows and 11 variables:

start

The start time of the first event, in ticks. In this video there are 30000 ticks per second.

duration

The length of the fragment, in ticks.

interval

The length of each event within the fragment, in ticks.

id

Trump or Clinton.

x

X coordinate of top-left corner of face.

y

Y coordinate of top-left corner of face.

width

Width of face.

height

Height of face.

row_id

ID of video frame.

key

Type of emotion measured.

value

Probability that the facial expression expresses this emotion.

Source

https://blog.exploratory.io/analyzing-emotions-using-facial-expressions-in-video-with-microsoft-ai-and-r-8f7585dd0780