World Bank indicators

A dataset containing both quantitative variables and qualitative variables for six countries from 2005 to 2017. The countries included are: Argentina, China, United Kingdom, Indonesia, Norway, and United States.

worldbank

Format

A data frame with 78 rows and 14 variables:

iso3c

Country code ISO3.

date

Year in which the data is for.

iso2c

Country code ISO2.

country

Name of the country.

perc_energy_fosfuel

Fossil fuel energy consumption ( % of total).

rnd_gdpshare

Research and development expenditure (% of GDP).

percgni_adj_gross_savings

Adjusted savings: gross savings (% of GNI).

real_netinc_percap

Adjusted net national income per capita (constant 2010 US$).

gdp_capita

GDP per capita (current USD).

top10perc_incshare

Income share held by highest 10 %.

employment_ratio

Employment to population ratio, 15+, total (%) (national estimate).

life_exp

Life expectancy at birth, total (years).

pop_growth

Population growth (annual %).

pop

Total population.

Source

https://databank.worldbank.org/home.aspx